requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
autoreject==0.4.3
black==23.11.0
blas==2.121
blas-devel==3.9.0
blis==0.7.11
bottleneck==1.3.7
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2024.2.2
catalogue==1.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.0
coverage==7.4.4
cycler==0.12.1
cymem==2.0.8
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
dora-search==0.1.12
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flake8==7.0.0
flashy==0.0.2
fonttools==4.50.0
freetype==2.12.1
fsspec==2023.10.0
ftfy==6.2.0
gitdb==4.0.11
gitpython==3.1.43
gmp==6.2.1
gmpy2==2.1.2
h5io==0.2.2
h5py==3.11.0
huggingface_hub==0.20.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jinja2==3.1.3
joblib==1.2.0
jpeg==9e
julius==0.2.7
jupyter_client==7.4.9
jupyter_core==5.5.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy-loader==0.3
lcms2==2.12
lerc==3.0
levenshtein==0.25.1
libblas==3.9.0
libcblas==3.9.0
libcxx==14.0.6
libdeflate==1.17
libffi==3.4.4
libgfortran==5.0.0
libgfortran5==13.2.0
liblapack==3.9.0
liblapacke==3.9.0
libopenblas==0.3.26
libpng==1.6.39
libprotobuf==3.20.3
libsodium==1.0.18
libtiff==4.5.1
libuv==1.44.2
libwebp-base==1.3.2
llvm-openmp==18.1.1
llvmlite==0.42.0
locate==1.1.1
lz4-c==1.9.4
markupsafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mne==1.6.1
mne-bids==0.14
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mypy==1.9.0
mypy_extensions==1.0.0
ncurses==6.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.10.2
ninja-base==1.10.2
numba==0.59.1
numexpr==2.8.7
numpy==1.26.4
numpy-base==1.26.4
omegaconf==2.3.0
openblas==0.3.26
openjpeg==2.3.0
openssl==3.2.1
osfclient==0.0.5
packaging==24.0
pandas==2.2.1
parso==0.8.4
pathspec==0.10.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
plac==1.1.3
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
preshed==3.0.9
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pygments==2.17.2
pymatreader==0.0.32
pyparsing==3.1.2
pysocks==1.7.1
pytest==8.1.1
pytest-cov==5.0.0
python==3.10.13
python-dateutil==2.8.2
python-levenshtein==0.25.1
python-tzdata==2023.3
pytorch==2.2.1
pytz==2023.3.post1
pyyaml==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
readline==8.2
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.12.0
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
sleef==3.5.1
smmap==5.0.1
spacy==2.3.9
sqlite==3.41.2
srsly==1.0.7
stack_data==0.6.2
submitit==1.5.1
sympy==1.12
thinc==7.4.6
threadpoolctl==2.2.0
tk==8.6.12
tokenizers==0.15.1
tomli==2.0.1
torchaudio==2.2.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
treetable==0.2.5
types-pyyaml==6.0.12.20240311
typing-extensions==4.9.0
typing_extensions==4.9.0
tzdata==2024a
urllib3==2.1.0
wandb==0.16.6
wasabi==0.10.1
wcwidth==0.2.13
wheel==0.41.2
wordfreq==3.1.1
xmltodict==0.13.0
xz==5.4.6
yaml==0.2.5
zeromq==4.3.5
zipp==3.17.0
zlib==1.2.13
zstd==1.5.5
